+++ cmd/healthcheck/loginapp/send_hello.py
@@ -3,6 +3,7 @@ import logging
 
 from enki import settings
 from enki.enkitype import AppAddr
+from enki.net import msgspec
 from enki.net.kbeclient import Client
 from enki.net.command.loginapp import HelloCommand
 from enki.misc import log
@@ -12,7 +13,7 @@ logger = logging.getLogger(__name__)
 async def main():
     log.setup_root_logger(logging.getLevelName(settings.LOG_LEVEL))
     try:
-        client = Client(settings.LOGIN_APP_ADDR)
+        client = Client(settings.LOGIN_APP_ADDR, msgspec.app.client.SPEC_BY_ID)
         res = await client.start()
         if not res.success:
             logger.error(f'Cannot connect to the "{settings.LOGIN_APP_ADDR}" server address '