+++ tests/utests/base.py
@@ -39,7 +39,7 @@ class EnkiBaseTestCase(asynctest.TestCase):
     def call_OnCreatedProxies(self):
         # entity_id = 2177 (Avatar)
         data = b'\xf8\x01\x13\x00\x00\x00\x07\x00\x95\x84\xfbb\x81\x08\x00\x00Avatar\x00'
-        msg_504, _ = MessageSerializer().deserialize(memoryview(data))
+        msg_504, _ = MessageSerializer(msgspec.app.client.SPEC_BY_ID).deserialize(memoryview(data))
         assert msg_504 is not None
         res_504 = OnCreatedProxiesHandler(self._entity_helper).handle(msg_504)
         assert res_504.success