+++ tests/utests/test_app/test_handlers/test_initSpaceData.py
@@ -1,5 +1,6 @@
 import unittest
 
+from enki.net import msgspec
 from enki.net import kbeclient
 
 from enki.app.handler.sdhandler import SpaceDataMgr, InitSpaceDataHandler
@@ -12,7 +13,7 @@ class InitSpaceDataTestCase(base.EnkiBaseTestCase):
 
     def test_ok(self):
         data = b'A\x00\x1f\x00\x01\x00\x00\x00_mapping\x00spaces/xinshoucun\x00'
-        msg, data_tail = kbeclient.MessageSerializer().deserialize(memoryview(data))
+        msg, data_tail = kbeclient.MessageSerializer(msgspec.app.client.SPEC_BY_ID).deserialize(memoryview(data))
         assert msg is not None, 'Invalid initial data'
 
         handler = InitSpaceDataHandler(SpaceDataMgr())