+++ tests/utests/test_app/test_handlers/test_onEntityEnterSpace.py
@@ -10,12 +10,12 @@ class OnEntityEnterSpaceTestCase(EnkiBaseTestCase):
     def test_ok(self):
         # Avatar id is 2199
         data = b'\xf8\x01\x13\x00\x00\x00\x07\x00\xf98\xfeb\x97\x08\x00\x00Avatar\x00'
-        msg_504, _ = MessageSerializer().deserialize(memoryview(data))
+        msg_504, _ = MessageSerializer(msgspec.app.client.SPEC_BY_ID).deserialize(memoryview(data))
         assert msg_504 is not None
         assert OnCreatedProxiesHandler(self._entity_helper).handle(msg_504).success
 
         data = b'\xfd\x01\t\x00\x97\x08\x00\x00\x01\x00\x00\x00\x00'
-        msg, data_tail = MessageSerializer().deserialize(memoryview(data))
+        msg, data_tail = MessageSerializer(msgspec.app.client.SPEC_BY_ID).deserialize(memoryview(data))
         assert msg is not None, 'Invalid initial data'
 
         handler = OnEntityEnterSpaceHandler(self._entity_helper)