+++ tests/utests/test_app/test_handlers/test_onStreamData.py
@@ -14,15 +14,15 @@ class onStreamDataTestCase(EnkiBaseTestCase):
 
     async def test_ok(self):
         data_514 = b'\x02\x02\x15\x00\x01\x00\x0e\x00\x00\x00unittest.data\x00\x01'
-        msg_514, data_tail = MessageSerializer().deserialize(memoryview(data_514))
+        msg_514, data_tail = MessageSerializer(msgspec.app.client.SPEC_BY_ID).deserialize(memoryview(data_514))
         assert msg_514 is not None, 'Invalid initial data'
 
         data_515 = b'\x03\x02\x14\x00\x01\x00\x0e\x00\x00\x00Unittest data\n'
-        msg_515, data_tail = MessageSerializer().deserialize(memoryview(data_515))
+        msg_515, data_tail = MessageSerializer(msgspec.app.client.SPEC_BY_ID).deserialize(memoryview(data_515))
         assert msg_515 is not None, 'Invalid initial data'
 
         data_516 = b'\x04\x02\x01\x00'
-        msg_516, data_tail = MessageSerializer().deserialize(memoryview(data_516))
+        msg_516, data_tail = MessageSerializer(msgspec.app.client.SPEC_BY_ID).deserialize(memoryview(data_516))
         assert msg_516 is not None, 'Invalid initial data'