+++ tests/utests/test_app/test_handlers/test_onUpdatePropertysOptimized.py
@@ -22,7 +22,7 @@ class OnUpdatePropertysOptimizedTestCase(base.EnkiBaseTestCase):
         handler = OnUpdatePropertysOptimizedHandler(self._app, self._entity_helper)
 
         data = b'\x0b\x00\x04\x00\x00\x00\x0e\x03\x0b\x00\x07\x00\x01\x00\t\x18\x00\x00\x00\x18\x00\t\x00\x01\x95\x9cDD\x14\xeaCD'
-        msg, data_tail = kbeclient.MessageSerializer().deserialize(memoryview(data))
+        msg, data_tail = kbeclient.MessageSerializer(msgspec.app.client.SPEC_BY_ID).deserialize(memoryview(data))
         assert msg is not None, 'Invalid initial data'
         result: HandlerResult = handler.handle(msg)
         assert result.success